feat: add features based on assignment #3

+++ src/main/java/com/saeed/waa/controllers/PostController.java
@@ -25,7 +25,7 @@ public class PostController {
     }
 
     @GetMapping("/{id}")
-    public PostDto getById(@PathVariable("id") int id) {
+    public PostDto getById(@PathVariable("id") Long id) {
         return postService.findById(id);
     }
